Committing changes to a lot of stuff
added more debug commands, as well as merged more of the UI
TODO: finalize weapon logic and fallbacks for weapons

+++ Assets/Scripts/Parker/Objects/BulletInfo.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class BulletInfo : MonoBehaviour 
 {
-	public int cost = 1;
+	public float cost = 1f;
 	public float fireRate = 1f;
 	public float travelSpeed = 10f;
 	public float travelAngle = 0f;
